% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-beetles.R
\docType{data}
\name{beetles}
\alias{beetles}
\alias{plants}
\alias{verges}
\title{Carabid beetles and vascular plants in Dutch roadside verges}
\format{
An object of class \code{data.frame} with 30 rows and 126 columns.
}
\source{
Raemakers, I.P., Schaffers, A.P., Sykora, V. and Heijerman, T. (2001)
The importance of plant communities in road verges as habitat for
insects. \emph{Proceedings of the Section Experimental and Applied
Entomology if the Netherlands Entomological Society} \strong{12}, 101--106.

ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence
Analysis: a new ordination method to relate two community
compositions. \emph{Ecology} \strong{85(3)}, 834--846
}
\usage{
data(beetles)
}
\description{
Counts of 126 beetle taxa and abundances of 173 vascular plant taxa
(expressed on the 1-9 van der Maarel scale) in 30 road-side verges in
the Netherlands.
}
\details{
This is the complete dataset of Raemakers et al (2001). ter Braak and
Schaffers (2004) only analyse a subset of 91 beetle taxa.
}
\examples{
data(beetles)
data(plants)
}
\keyword{datasets}
