% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bryophyte.R
\docType{data}
\name{bryophyte}
\alias{bryophyte}
\alias{vascular}
\title{Bryophytes and vascular plants in Carpathian spring meadows}
\format{
An object of class \code{data.frame} with 70 rows and 30 columns.
}
\source{
Hajek, M., Hekera, P. and Hajkova, P. (2002) Spring fen vegetations
and water chemistry in the Western Carpathian flysch zone. \emph{Folia
geobotanica} \strong{37}, 205--224

ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence
Analysis: a new ordination method to relate two community
compositions. \emph{Ecology} \strong{85(3)}, 834--846
}
\usage{
data(bryophyte)
}
\description{
The data consist of observations on 30 bryophyte and 123 vascular
plant species in 70 spring meadows (sites, samples). The species data
are a subset of only those species occurring in five or more meadows.
}
\examples{
data(bryophyte)
data(vascular)
}
\keyword{datasets}
