% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corAxis.R
\name{corAxis}
\alias{corAxis}
\alias{corAxis.default}
\alias{corAxis.symcoca}
\title{Correlation between ordination axes}
\usage{
corAxis(x, ...)

\method{corAxis}{default}(x, ...)

\method{corAxis}{symcoca}(x, axes = NULL, ...)
}
\arguments{
\item{x}{an ordination object. Only methods for objects of class
\link{symcoca} are currently available.}

\item{...}{arguments to be passed on to other methods.}

\item{axes}{numeric; the number of axes to calculate the correlation
coefficients for. If \code{NULL}, coefficients for all axes are
returned.}
}
\value{
A named vector containing the correlation coefficients for the
requested axes.
}
\description{
Calculates the Pearson product-moment correlation coefficient for the
site scores of ordination axes.
}
\note{
The arguments for \code{cor} are hard coded at their
defaults, see \link{cor} for details. A more flexible version
is planned that will allow arguments to be passed to \code{cor}.
}
\examples{
\dontshow{od <- options(digits = 4)}
## load some data
data(beetles)
data(plants)

## log transform the beetle data
beetles <- log(beetles + 1)

## symmetric Co-CA model
beetles.sym <- coca(beetles ~ ., data = plants, method = "symmetric")

## correlations between axes
corAxis(beetles.sym)
\dontshow{options(od)}
}
\seealso{
\link{cor}, for the main analysis function.
}
\author{
Gavin L. Simpson
}
\keyword{univar}
