% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfit.coca.R
\name{envfit.coca}
\alias{envfit.coca}
\title{Fit an environmental vector or factor to a Co-CA ordination}
\usage{
\method{envfit}{coca}(
  ord,
  env,
  which = c("response", "predictor"),
  choices = c(1, 2),
  scaling = FALSE,
  w,
  na.rm = FALSE,
  strata = NULL,
  permutations = 999,
  ...
)
}
\arguments{
\item{ord}{a Co-CA ordination object, the result of a call to
\link{coca}.}

\item{env}{a data frame, matrix or vector of environmental/external
variable(s) to be fitted to the ordination. The variables may be
of a mixed type (factors and continuous variables) in a data frame.}

\item{which}{character; which of the response or predictor ordinations
should be used during fitting of vectors and factors.}

\item{choices}{numeric; the axes to which vectors and factors are
fitted.}

\item{scaling}{logical; should scaling be applied. See
\link{scores.symcoca}.}

\item{w}{weights used in fitting vectors and factors.}

\item{na.rm}{Remove points with missing values in ordination scores or
environmental variables. The operation is case-wise; the whole row of
data is removed if there is a missing value and \code{na.rm = TRUE}.}

\item{strata}{An integer vector or factor specifying the strata for
permutation. If supplied, observations are permuted only within the
specified strata.}

\item{permutations}{Number of permutations for assessing significance
of vectors or factors. Set to \code{0} to skip permutations.}

\item{...}{Arguments passed to \link[vegan:vectorfit]{vectorfit} and
\link[vegan:factorfit]{factorfit}.}
}
\value{
Returns an object of class \link[vegan:envfit]{envfit}.
}
\description{
The function fits environmental vectors or factors to a Co-CA
ordination. The projections of points onto vectors have maximum
correlation with corresponding environmental variables, and the
factors show the averages of factor levels.
}
\details{
See \link[vegan:envfit]{envfit} for details of the method.
}
\section{Warning}{

\%   This function is experimental. The vectors fitted are unweighted
\%   whereas to be consistent with the \link[vegan:cca]{cca} method of
\%   \link[vegan:envfit]{envfit} they should be weighted. I just need to work out
\%   the correct sets of weights to use.
\%
}

\examples{
## symmetric CoCA
data(beetles)
data(plants)

## log transform the bettle data
beetles <- log(beetles + 1)

## fit the model
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")

## load the environmental data
data(verges)

## fit vectors for the environmental data
sol <- vegan::envfit(bp.sym, verges, which = "response")
\donttest{sol}

## plot the response matrix and the fitted vectors
biplot(bp.sym, which = "y1")
plot(sol)
}
\seealso{
\link{coca} for fitting models. \link[vegan:envfit]{envfit} for
details of the generic function and the computations performed.
}
\author{
Gavin L. Simpson. The code interfaces with and uses code from
\link[vegan:envfit]{envfit} for the main computations, which was written by
Jari Oksanen.
}
\keyword{htest}
