% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.symcoca.R, R/print.fitted.symcoca.R
\name{fitted.symcoca}
\alias{fitted.symcoca}
\alias{print.fitted.symcoca}
\title{Fitted values of a Symmetric Co-Correspondence analysis model.}
\usage{
\method{fitted}{symcoca}(object, which = c("y1", "y2"), ...)

\method{print}{fitted.symcoca}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object, x}{an object of class \code{"symcoca"}}

\item{which}{character; should the response or predictor scores be
plotted. Can be specified in several ways: \emph{response} choices
are one from \code{c("y", "Y", "y1", "response")}; \emph{predictor}
choices are one from \code{c("x", "X", "y2", "predictor")}.}

\item{...}{arguments to be passed to other methods.}

\item{digits}{Number of significant digits for printing.}
}
\value{
A list with the following components:
\itemize{
\item \strong{Y}: the fitted values for the "response" matrix.
\item \strong{X}: the fitted values for the "predictor" matrix.
\item \strong{nam.dat}: a vector containing the names of the "response" and "predictor" matrices respectively. Used for printing the results.
}
}
\description{
Calculates and extracts the fitted values of a Symmetric
Co-Correspondence analysis model.
}
\note{
This function needs an update and to allow option to restrict
fitted values to specified axes, and the names of the returned objects
need making more obvious!
}
\examples{
## symmetric CoCA
data(beetles)
data(plants)

## log transform the bettle data
beetles <- log(beetles + 1)

## fit the model
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")

## fitted values
bp.fit <- fitted(bp.sym)
\donttest{bp.fit}

## fitted values for beetles only
beetle.fit <- fitted(bp.sym, which = "y1")
}
\references{
Ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence
Analysis: a new ordination method to relate two community
compositions. \emph{Ecology} \strong{85(3)}, 834--846
}
\seealso{
The model fitting function \link{coca}
}
\author{
Gavin L. Simpson, based on Matlab code by C.J.F. ter Braak and
A.P. Schaffers.
}
\keyword{multivariate}
