% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.coca.R
\name{points.coca}
\alias{points.coca}
\title{Add points to a Co-CA plot}
\usage{
\method{points}{coca}(
  x,
  display = c("sites", "species"),
  which = c("response", "predictor"),
  choices = c(1, 2),
  scaling = FALSE,
  select,
  ...
)
}
\arguments{
\item{x}{an object inheriting from class \link{coca}.}

\item{display}{character; one of the stated choices. Indicates which
scores to use to draw points.}

\item{which}{character; one of the stated choices. Indicates which of
the response or predictor data sets is used to select scores from.}

\item{choices}{The Co-CA axes to draw points for.}

\item{scaling}{logical; should species scores in a symmetric Co-CA be
rescaled?}

\item{select}{Items to be displayed. This can either be a logical
vector which is \code{TRUE} for displayed items or a vector of
indices of displayed items.}

\item{...}{Arguments passed to other methods}
}
\value{
Returns the plotted x and y coordinates as a matrix.
}
\description{
Draws points on the current graphic device based on in
supplied \link{coca} model object.
}
\details{
The visual appearance of the plotted points can be controlled by
supplying appropriate graphical parameters via the \code{...}
argument. See \link{par} for details.
}
\examples{
## symmetric CoCA
data(beetles)
data(plants)
     
## log transform the bettle data
beetles <- log(beetles + 1)
     
## fit the model
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")

## draw a plot of the response scores
plot(bp.sym, type = "none")
points(bp.sym, display = "sites", col = "blue", pch = 16)
points(bp.sym, display = "species", col = "red", pch = 3, cex = 0.8)
}
\seealso{
\link{plot} methods; \link{plot.predcoca} and
\link{plot.symcoca}.
}
\author{
Gavin L. Simpson
}
\keyword{methods}
