% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.symcoca.R
\name{resid.symcoca}
\alias{resid.symcoca}
\alias{residuals.symcoca}
\title{Extract Model Residuals}
\usage{
\method{resid}{symcoca}(object, ...)

\method{residuals}{symcoca}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"symcoca"}.}

\item{...}{arguments to be passed to other methods.}
}
\value{
A list containing the residuals for the response and the
predictor with the following components:
\itemize{
\item \strong{Y}: residuals of the fit to the response.
\item \strong{X}: residuals of the fit to the predictor.
}
}
\description{
Extracts the residuals of the fitted model of a symmetric CoCA
to the response and the predictor.
}
\examples{
\dontshow{od <- options(digits = 4)}
data(bryophyte)
data(vascular)
bryo.sym <- coca(bryophyte ~ ., data = vascular, method = "symmetric")
eps <- resid(bryo.sym)
\dontshow{options(od)}
}
\seealso{
\link{symcoca}
}
\author{
Gavin L. Simpson
}
\keyword{multivariate}
