% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights.symcoca}
\alias{weights.symcoca}
\alias{weights.predcoca}
\title{Extract CoCA model weights}
\usage{
\method{weights}{symcoca}(object, ...)

\method{weights}{predcoca}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"symcoca"} or \code{"predcoca"}.}

\item{...}{arguments passed to other methods. Not used.}
}
\value{
A numeric vector of common site weights is currently returned. These
correspond to \eqn{R_{0}}{R[0]} in ter Braak and Schaffers (2004).
}
\description{
Extractor function to identify and select appropriate analysis weights
from Co-correspondence Analysis models.
}
\examples{
\dontshow{od <- options(digits = 4)}
## symmetric CoCA
data(beetles)
data(plants)

## log transform the bettle data
beetles <- log(beetles + 1)

## fit the model
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")

## weights == R[0]
weights(bp.sym)
\dontshow{options(od)}
}
\references{
ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence
Analysis: a new ordination method to relate two community
compositions. \emph{Ecology} \strong{85(3)}, 834--846.
}
\seealso{
\link{envfit.coca} uses these weights to estimate
weighted correlations in ordination space. See \link{coca} for
details on creating CoCA models.
}
\author{
Gavin L. Simpson
}
\keyword{utilities}
