\name{bryophyte}
\alias{bryophyte}
\alias{vascular}
\docType{data}
\title{Bryophytes and vascular plants in Carpathian spring meadows}
\description{
  The data consist of observations on 30 bryophyte and 123 vascular
  plant species in 70 spring meadows (sites, samples). The species data
  are a subset of only those species occuring in five or more meadows.
}
\usage{
data(bryophyte)
data(vascular)}

\source{

  Hajek, M., Hekera, P. and Hajkova, P. (2002) Spring fen vegetations
  and water chemistry in the Western Carpathian flysch zone. \emph{Folia
  geobotanica} \bold{37}, 205--224
  
  ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence
  Analysis: a new ordination method to relate two community
  compositions. \emph{Ecology} \bold{85(3)}, 834--846

}
\examples{
data(bryophyte)
data(vascular)
}
\keyword{datasets}
