% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocorresp-package.R
\docType{package}
\name{cocorresp-package}
\alias{cocorresp-package}
\alias{cocorresp}
\title{Co-correspondence analysis ordination methods for community ecology}
\description{
Fits predictive and symmetric co-correspondence analysis (CoCA) models
to relate one data matrix to another data matrix. More specifically,
CoCA maximises the weighted covariance between the weighted averaged
species scores of one community and the weighted averaged species
scores of another community. CoCA attempts to find patterns that are
common to both communities.
}
\details{
The main interface function is \link{coca} which accepts a
formula or two community data matrices. An appropriate formula is
\verb{Y ~ ., data = X} and the associated \code{data} object from
which \code{.} will be looked up. The \code{method} argument is used
to select from the two forms of CoCA: \code{method = "predictive"} for
predictive CoCA (the default), and \code{method = "symmetric"} for
symmetric CoCA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gavinsimpson.github.io/cocorresp/}
  \item Report bugs at \url{https://github.com/gavinsimpson/cocorresp/issues}
}

}
\author{
Original Matlab routines by C.J.F. ter Braak and A.P. Schaffers. R port
by Gavin L. Simpson. Function simpls based on \code{simpls.fit} (package
pls) by Ron Wehrens and Bjorn-Helge Mevik.

Maintainer: Gavin L. Simpson \href{mailto:ucfagls@gmail.com}{ucfagls@gmail.com}
}
\keyword{package}
