\name{cocorresp-package}
\alias{cocorresp-package}
\alias{cocorresp}
\docType{package}
\title{
Co-correspondence analysis ordination methods for community ecology
}
\description{
  Fits predictive and symmetric co-correspondence analysis (CoCA) models
  to relate one data matrix to another data matrix. More specifically,
  CoCA maximises the weighted covariance between the weighted averaged
  species scores of one community and the weighted averaged species
  scores of another community. CoCA attempts to find patterns that are
  common to both communitities.
}
\details{
\tabular{ll}{
Package: \tab cocorresp\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2013-12-12\cr
Depends: \tab vegan\cr
License: \tab GPL Version 2\cr
Encoding: \tab UTF-8\cr
}
}
\author{
Original Matlab routines by C.J.F. ter Braak and A.P. Schaffers. R port
by Gavin L. Simpson. Function simpls based on simpls.fit (package pls)
by Ron Wehrens and Bjorn-Helge Mevik.

Maintainer: Gavin L. Simpson <gavin.simpson@ucl.ac.uk>
}
\keyword{ package }
