% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCoinertia.R, R/print.coinertia.R,
%   R/print.summary.coinertia.R, R/summary.coinertia.R
\name{fitCoinertia}
\alias{fitCoinertia}
\alias{print.fitCoinertia}
\alias{print.summary.fitCoinertia}
\alias{summary.fitCoinertia}
\title{Co-inertia analysis}
\usage{
fitCoinertia(X, Dp, Y, Dq, Dn, n.axes)

\method{print}{fitCoinertia}(
  x,
  axes = c(1:min(6, x$n.axes)),
  digits = max(3, getOption("digits") - 3),
  ...
)

\method{print}{summary.fitCoinertia}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{fitCoinertia}(object, ...)
}
\arguments{
\item{X}{\eqn{Q_1}, matrix of expected abundances under row-column
independence in the original \code{Y} species matrix when
treated as a contingency table.}

\item{Dp}{\eqn{K_1}, species (column) weights for \code{X}.}

\item{Y}{\eqn{Q_2}, matrix of expected abundances under row-column
independence in the original \code{X} species matrix when
treated as a contingency table.}

\item{Dq}{\eqn{K_2}, species (column) weights for \code{Y}.}

\item{Dn}{site weights \eqn{R_0}.}

\item{n.axes}{number of axes to calculate the co-inertia analysis
for.}

\item{axes}{the number of axes to display when printing.}

\item{digits}{the number of significant digits to use in printing}

\item{object}{an object of class \code{coinertia}.}

\item{\ldots}{arguments passed to other functions. Currently ignored.}
}
\value{
A list with the following components:
\item{U1 }{column weights of \code{X}.}
\item{U2 }{column weights of \code{Y}.}
\item{X1 }{rowscores of \code{X}.}
\item{X2 }{rowscores of \code{Y}.}
\item{lambda }{the Eigenvalues (squares of the singular values).}
\item{n.axes }{number of axes requested.}
\item{call }{the matched function call.}
}
\description{
Performs a co-inertia of the triplets \eqn{(Q_1,K_1,R_0)} and
\eqn{(Q_2,K_2,R_0)}.
}
\note{
This function is not yet meant to be called directly by the
user. If you wish to use it directly, see the function definition for
\link{symcoca} which demonstrates how to prepare the
relevant input matrices.

Note that in this function, \code{X} corresponds to the input
matrix \code{y} and \code{Y} corresponds to the
input matrix \code{x} in \link{symcoca}.
Confusing! This will be changed in a future release but for now the
arguments follow those of the original Matlab code - perhaps a little
too closely!
}
\references{
Ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence
Analysis: a new ordination method to relate two community
compositions. \emph{Ecology} \strong{85(3)}, 834--846

Doledec, S and Chessel, D. (1994) Co-inertia analysis: a method for
studying species-environment relationships. \emph{Freshwater Biology}
\strong{31}, 277--294
}
\seealso{
\link{symcoca} for the function that calls
\code{fitCoinertia} and \link{coinertiaI} for co-inertia analysis
using identity matrices for \eqn{K_1}, \eqn{K_2}, and \eqn{R_0}
}
\author{
Original Matlab code by C.J.F. ter Braak and A.P. Schaffers. R
port by Gavin L. Simpson.
}
\keyword{models}
