% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R, R/rescale.default.R,
%   R/rescale.symcoca.R
\name{rescale}
\alias{rescale}
\alias{rescale.default}
\alias{rescale.symcoca}
\title{Rescales CoCA species scores}
\usage{
rescale(object, ...)

\method{rescale}{default}(object, ...)

\method{rescale}{symcoca}(object, choices = NULL, display = c("species", "sites"), ...)
}
\arguments{
\item{object}{an R object. Currently only objects of class
\code{"symcoca"} are supported.}

\item{...}{other arguments to be passed to \code{rescale}
methods. Currently not used.}

\item{choices}{numeric; which axes to rescale and return.}

\item{display}{character; which type of scores to rescale.}
}
\value{
Returns a list with one or both of the following components:
\item{species}{rescaled species scores for the response}
\item{sites}{rescaled species scores for the predictor}
}
\description{
Rescales CoCA species scores to the quarter root of the
eigenvalues.
}
\details{
Currently only implemented for objects of class \code{"symcoca"}.
}
\examples{
\dontshow{od <- options(digits = 4)}
data(bryophyte)
data(vascular)

bryo.sym <- coca(bryophyte ~ ., data = vascular, method = "symmetric")

\donttest{rescale(bryo.sym, axes = 1:2)}
\dontshow{options(od)}
}
\seealso{
\link{symcoca}
}
\author{
Matlab original by C.J.F. ter Braak and A.P. Schaffers.
R port by Gavin L. Simpson.
}
\keyword{multivariate}
