% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.predcoca.R, R/scores.symcoca.R
\name{scores.predcoca}
\alias{scores.predcoca}
\alias{scores-methods}
\alias{scores.symcoca}
\title{Get Species or Site Scores from an Ordination}
\usage{
\method{scores}{predcoca}(x, choices = c(1, 2), display = c("sites", "species"), ...)

\method{scores}{symcoca}(
  x,
  choices = c(1, 2),
  display = c("sites", "species"),
  scaling = FALSE,
  ...
)
}
\arguments{
\item{x}{an ordination result}

\item{choices}{numeric; the ordination axes to return.}

\item{display}{partial match to access scores for "sites"
"species", "loadings" or "xmatrix". The latter
two are only available for \link{symcoca}.}

\item{...}{arguments to be passed to other methods.}

\item{scaling}{logical; whether scores should be rescaled
by the quarter root of the eigenvalues using
\link{rescale.symcoca}.}
}
\value{
A list with one or more components containing matrices of the
requested scores:
\itemize{
\item \strong{species}: A list with two components, \code{Y} and \code{X}, containing the species scores for the response matrix \code{Y} and the predictor matrix \code{X} respectively.
\item \strong{sites}: A list with two components, \code{Y} and \code{X}, containing the site scores for the response matrix \code{Y} and the predictor matrix \code{X} respectively.
\item \strong{loadings}: A list with two components, \code{Y} and \code{X} containing the loadings for the response and predictor matrix. For \link{symcoca} only.
\item \strong{xmatrix}: The X matrix. For \link{symcoca} only.
}
}
\description{
Function to access either species or site scores for specified axes in
co-correspondence analysis ordination methods.
}
\details{
Implements a \link[vegan:scores]{scores} method for symmetric
co-correspondence analysis ordination results.
}
\examples{
\dontshow{od <- options(digits = 4)}
## load some data
data(beetles)
data(plants)

## log transform the bettle data
beetles <- log(beetles + 1)

## fit the model, a symmetric CoCA
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")

## extract the scores
scr <- scores(bp.sym)

## predictive CoCA using SIMPLS and formula interface
bp.pred <- coca(beetles ~ ., data = plants)
scr2 <- scores(bp.pred)

\dontshow{options(od)}
}
\references{
ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence
Analysis: a new ordination method to relate two community
compositions. \emph{Ecology} \strong{85(3)}, 834--846
}
\seealso{
\link[vegan:scores]{scores}, for further details on the method.
}
\author{
Gavin L. Simpson, based on Matlab code by C.J.F. ter Braak and
A.P. Schaffers.
}
\keyword{methods}
