% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.predcoca.R,
%   R/print.summary.symcoca.R, R/summary.predcoca.R, R/summary.symcoca.R
\name{print.summary.predcoca}
\alias{print.summary.predcoca}
\alias{summary-methods}
\alias{print.summary.symcoca}
\alias{summary.predcoca}
\alias{summary.symcoca}
\title{Summarizing Co-CA Model Fits}
\usage{
\method{print}{summary.predcoca}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{summary.symcoca}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{predcoca}(object, axes = NULL, ...)

\method{summary}{symcoca}(object, ...)
}
\arguments{
\item{digits}{Number of significant digits for printing.}

\item{...}{arguments to be passed to other methods.}

\item{object, x}{an object of class \code{"predcoca"} or
\code{"symcoca"}. Generally the result of a call to
\link{coca}.}

\item{axes}{numeric; how many axes to summarise? The default is to
display 6 axes or all available axes, whichever is the smaller.}
}
\value{
A list with the some of the following components:
\itemize{
\item \strong{cocaScores}: The site and/or species scores for the axes requested.
\item \strong{call}: The call used to fit the model.
\item \strong{lambda}: The eigenvalues for the axes requested. Not for \link{predcoca.simpls}.
\item \strong{namY, namX}: the names of the response and predictor either supplied by the user or derived from the original call.
\item \strong{loadings}: a list with two components \code{loadings1} and \code{loadings2}, which refer to the response and the predictor matrices respectively. (Only for predictive CoCA models.)
\item \strong{varianceExp}: a list with components \code{Yblock} and \code{Xblock} containing the amount of variance explained on each CoCA axis in the response and the predictor respectively. (Only for predictive CoCA models.)
\item \strong{totalVar}: a list with components \code{Yblock} and \code{Xblock} containing the total variance in the response and the predictor data sets respectively
\item \strong{inertia}: a list with components \code{total} and \code{residual} containing the total and residual inertia (variance) in the response and the predictor matrices of a symmetric CoCA model. (Only for symmetric CoCA models.)
\item \strong{scaling}: the scaling used/requested. (Only for symmetric CoCA models.)
}
}
\description{
\code{summary} methods for classes \code{"predcoca"} and
\code{"symcoca"}. These provide a summary of the main results of a
Co-Correspondence Analysis model.
}
\examples{
\dontshow{od <- options(digits = 4)}
## symmetric CoCA
data(beetles)
data(plants)

## log transform the bettle data
beetles <- log(beetles + 1)

## fit the model
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")
summary(bp.sym)

## Predictive CoCA
bp.pred <- coca(beetles ~ ., data = plants)
summary(bp.pred, axes = 1:2)
\dontshow{options(od)}
}
\seealso{
The model fitting function \link{coca}
}
\author{
Gavin L. Simpson
}
\keyword{multivariate}
